Remove bandit lint tool
I'm not convinced this has demonstrated its benefit to the code base.
When introduced, it changed one assert usage to a raise -- that's it --
which doesn't seem all that interesting of a change. Beyond that, this
tool requires many useless "nosec" comments for each suprocess call,
which I find distracting.

The tool itself adds overhead -- even if small -- to all contributor's
test runs and CI. IMO, we're better off without it.